Keep all delta classes together
