+++ src/CMakeLists.txt
@@ -43,8 +43,6 @@ add_library(libmata STATIC
 	strings/nfa-segmentation.cc
 	strings/nfa-strings.cc
 	nfa/delta.cc
-	nfa/post.cc
-	nfa/move.cc
 	nfa/operations.cc
 )
